# theslotbot/apps/api/openapi/spec.yaml
openapi: 3.1.0
info:
  title: theslotbot API
  version: 1.0.0
  description: |
    WhatsApp booking automation backend for theslotbot.

    ## Contract rules
    - This file is the source of truth for the API shape.
    - Only Dev 1 merges changes to this file (after team discussion).
    - Dev 3 runs `msw` against this spec from Day 1 and is never blocked.
    - Any drift between this spec and the real route handlers is a bug,
      caught at Gate 1 (Day 10) integration checkpoint.
    - All monetary values are in **paise** (integer). 100 paise = ₹1.
    - All timestamps are ISO 8601 with timezone offset.
    - All IDs are UUIDs (string).
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api-staging.theslotbot.com/api
    description: Staging
  - url: https://api.theslotbot.com/api
    description: Production

# ─────────────────────────────────────────────
# SECURITY
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT. Obtain by calling Supabase Auth signInWithPassword()
        from the Admin Panel. Pass the `access_token` as the Bearer token.
        The API validates this against SUPABASE_URL + SUPABASE_ANON_KEY.

  # ─── Reusable Schemas ─────────────────────────────────────────────
  schemas:
...
    BookingStatus:
      type: string
      enum:
        - pending_confirmation
        - confirmed
        - completed
        - cancelled
        - no_show
    BookingSource:
      type: string
      enum:
        - whatsapp
        - admin_manual
    ServiceCategory:
      type: string
      enum:
        - short_cycle
        - medium_cycle
        - long_cycle
...
    RevisitCampaignStatus:
      type: string
      enum:
        - none
        - day30_sent
        - day37_sent
        - converted
        - non_responder
...
    Service:
      type: object
      required: [id, salonId, name, price, durationMinutes, category, isActive, displayOrder]
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        name:
          type: string
          example: HydraFacial
        description:
          type: string
          nullable: true
        price:
          type: integer
          description: In paise. 350000 = ₹3,500.
          example: 350000
        durationMinutes:
          type: integer
          example: 60
        category:
          $ref: '#/components/schemas/ServiceCategory'
        isActive:
          type: boolean
        displayOrder:
          type: integer
    Slot:
      type: object
      required: [id, salonId, startTime, endTime, isBlocked]
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isBlocked:
          type: boolean
        stylistId:
          type: string
          format: uuid
          nullable: true
        bookingId:
          type: string
          format: uuid
          nullable: true
          description: Null when available. Set atomically on booking claim.
    Customer:
      type: object
      required: [id, salonId, phoneNumber, name, whatsappOptIn, revisitCampaignStatus, nonResponderCount, isNumberInvalid, createdAt]
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        phoneNumber:
          type: string
          example: "+919876543210"
        name:
          type: string
        whatsappOptIn:
          type: boolean
          description: |
            Explicit opt-in for marketing-category messages.
            Default false. Set true only when customer replies YES
            during the booking confirmation opt-in step. (Risk D4 fix)
        lastVisitDate:
          type: string
          format: date
          nullable: true
          description: |
            The date of the actual appointment (slotStart::date),
            NOT when staff clicked "mark complete". (Risk D1 fix)
        lastVisitServiceId:
          type: string
          format: uuid
          nullable: true
        revisitCampaignStatus:
          $ref: '#/components/schemas/RevisitCampaignStatus'
        nonResponderCount:
          type: integer
        isNumberInvalid:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Booking:
      type: object
      required: [id, salonId, customerId, serviceId, status, source, slotStart, slotEnd, actualVisitDate, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BookingStatus'
        source:
          $ref: '#/components/schemas/BookingSource'
        slotStart:
          type: string
          format: date-time
        slotEnd:
          type: string
          format: date-time
        actualVisitDate:
          type: string
          format: date
          description: |
            Set from slotStart::date at creation time. Used by campaign
            workers, not completedAt. Never changes after creation. (Risk D1)
        reminder24hSent:
          type: boolean
        reminder3hSent:
          type: boolean
        reviewRequestSent:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        completedById:
          type: string
          format: uuid
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        # Expanded relations (included when ?include=customer,service)
        customer:
          $ref: '#/components/schemas/Customer'
          description: Populated when include=customer is requested
        service:
          $ref: '#/components/schemas/Service'
          description: Populated when include=service is requested
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code from shared/constants ERROR_CODES
              example: SLOT_ALREADY_CLAIMED
            message:
              type: string
              description: Human-readable message (safe to display)
            details:
              description: Additional structured context (validation errors, etc.)

    # ── Pagination ────────────────────────────────────────────────────
    PaginatedMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer

    # ── Campaign Report ───────────────────────────────────────────────
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    DateFromParam:
      name: dateFrom
      in: query
      schema:
        type: string
        format: date
      description: Filter from this date (inclusive), in salon local timezone.
    DateToParam:
      name: dateTo
      in: query
      schema:
        type: string
        format: date
      description: Filter to this date (inclusive), in salon local timezone.

  # ── Common Responses ───────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - field: slotId
                  message: Required

# ─────────────────────────────────────────────
# GLOBAL SECURITY
# All routes require BearerAuth unless marked with security: []
# ─────────────────────────────────────────────
security:
  - BearerAuth: []

# ─────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────
paths:
...
  /services:
    get:
      summary: List all active services for the salon
      description: Returns services ordered by displayOrder. Inactive services excluded.
      tags: [Services]
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /slots:
    get:
      summary: List available slots for a given date and service
      description: |
        Returns open, unblocked slots that are long enough for the requested
        service duration. Used by the Admin Panel's slot calendar view.
        The WhatsApp state machine calls this internally — it is not called
        directly by the customer.

        Overlap check (Risk B1) is applied server-side: only slots with no
        conflicting booking in the (start, start+duration+buffer) window
        are returned.
      tags: [Slots]
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD, interpreted in salon timezone
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Available slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Slot'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /slots/{slotId}/block:
    parameters:
      - name: slotId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Block or unblock a slot (reception and above)
      description: |
        Used by reception to block specific times (staff breaks, holidays).
        Cannot block a slot that already has a confirmed booking.
      tags: [Slots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isBlocked]
              properties:
                isBlocked:
                  type: boolean
      responses:
        '200':
          description: Slot updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/Slot'
        '409':
          description: Cannot block a slot with an active booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: BOOKING_NOT_CANCELLABLE
                  message: This slot has a confirmed booking and cannot be blocked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════
  # BOOKINGS
  # ════════════════════════════════════════════
  /bookings:
    get:
      summary: List bookings with filters
      description: |
        Primary endpoint for the Admin Panel's bookings views.
        Supports filtering by date range, status, and customer.
        Results include expanded customer and service by default.
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/DateFromParam'
        - $ref: '#/components/parameters/DateToParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: today
          in: query
          schema:
            type: boolean
          description: Shorthand for dateFrom=today&dateTo=today (salon timezone)
      responses:
        '200':
          description: Paginated bookings list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
                      meta:
                        $ref: '#/components/schemas/PaginatedMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a booking manually (admin walk-in logging)
      description: |
        Used by reception staff to log walk-in visits or create bookings
        on behalf of a customer who called in.
        Source is set to admin_manual.

        The same overlap check (Risk B1) applies as for WhatsApp bookings.
        Walk-in creation resets the customer's revisit campaign state to
        none and updates lastVisitDate. (Risk D2 fix)
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerPhone, customerName, serviceId, slotId]
              properties:
                customerPhone:
                  type: string
                  description: E.164 format. Upserts customer record.
                  example: "+919876543210"
                customerName:
                  type: string
                serviceId:
                  type: string
                  format: uuid
                slotId:
                  type: string
                  format: uuid
                notes:
                  type: string
                markCompleteImmediately:
                  type: boolean
                  default: false
                  description: |
                    Set true for walk-ins logged retrospectively.
                    Immediately sets status=completed and updates
                    lastVisitDate. Triggers review request job.
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/Booking'
        '409':
          description: Slot already claimed or overlap detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                slotClaimed:
                  value:
                    success: false
                    error:
                      code: SLOT_ALREADY_CLAIMED
                      message: This slot was just taken. Please choose another.
                overlapDetected:
                  value:
                    success: false
                    error:
                      code: SLOT_OVERLAP_DETECTED
                      message: This time window conflicts with an existing booking.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /customers:
    get:
      summary: Search customers
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: search
          in: query
          schema:
            type: string
          description: Searches name and phoneNumber (partial match)
        - name: campaignStatus
          in: query
          schema:
            $ref: '#/components/schemas/RevisitCampaignStatus'
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
                      meta:
                        $ref: '#/components/schemas/PaginatedMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
...